# Add channel handler for server side pipelining.