--- a/IntegrationTests/tests_01_http/test_13_http_pipelining.sh
+++ b/IntegrationTests/tests_01_http/test_13_http_pipelining.sh
@@ -28,77 +28,53 @@ echo -e 'GET /dynamic/count-to-ten HTTP/1.1\r\n\r\nGET /dynamic/count-to-ten HTT
 backslash_r=$(echo -ne '\r')
 cat > "$tmp/expected" <<EOF
 HTTP/1.1 200 OK$backslash_r
-Transfer-Encoding: chunked$backslash_r
-Connection: Keep-Alive$backslash_r
-Content-Type: text/plain$backslash_r
+transfer-encoding: chunked$backslash_r
 $backslash_r
+1$backslash_r
+1$backslash_r
+1$backslash_r
 2$backslash_r
-1
-$backslash_r
-2$backslash_r
-2
-$backslash_r
-2$backslash_r
-3
-$backslash_r
-2$backslash_r
-4
-$backslash_r
-2$backslash_r
-5
-$backslash_r
-2$backslash_r
-6
-$backslash_r
-2$backslash_r
-7
-$backslash_r
-2$backslash_r
-8
-$backslash_r
-2$backslash_r
-9
-$backslash_r
+1$backslash_r
 3$backslash_r
-10
-$backslash_r
+1$backslash_r
+4$backslash_r
+1$backslash_r
+5$backslash_r
+1$backslash_r
+6$backslash_r
+1$backslash_r
+7$backslash_r
+1$backslash_r
+8$backslash_r
+1$backslash_r
+9$backslash_r
+2$backslash_r
+10$backslash_r
 0$backslash_r
 $backslash_r
 HTTP/1.1 200 OK$backslash_r
-Transfer-Encoding: chunked$backslash_r
-Connection: Keep-Alive$backslash_r
-Content-Type: text/plain$backslash_r
+transfer-encoding: chunked$backslash_r
 $backslash_r
+1$backslash_r
+1$backslash_r
+1$backslash_r
 2$backslash_r
-1
-$backslash_r
-2$backslash_r
-2
-$backslash_r
-2$backslash_r
-3
-$backslash_r
-2$backslash_r
-4
-$backslash_r
-2$backslash_r
-5
-$backslash_r
-2$backslash_r
-6
-$backslash_r
-2$backslash_r
-7
-$backslash_r
-2$backslash_r
-8
-$backslash_r
-2$backslash_r
-9
-$backslash_r
+1$backslash_r
 3$backslash_r
-10
-$backslash_r
+1$backslash_r
+4$backslash_r
+1$backslash_r
+5$backslash_r
+1$backslash_r
+6$backslash_r
+1$backslash_r
+7$backslash_r
+1$backslash_r
+8$backslash_r
+1$backslash_r
+9$backslash_r
+2$backslash_r
+10$backslash_r
 0$backslash_r
 $backslash_r
 EOF