--- a/Tests/LinuxMain.swift
+++ b/Tests/LinuxMain.swift
@@ -52,6 +52,7 @@ import XCTest
          testCase(HTTPResponseCompressorTest.allTests),
          testCase(HTTPResponseEncoderTests.allTests),
          testCase(HTTPServerClientTest.allTests),
+         testCase(HTTPServerPipelineHandlerTest.allTests),
          testCase(HTTPTest.allTests),
          testCase(HTTPUpgradeTestCase.allTests),
          testCase(HappyEyeballsTest.allTests),