@normanmaurer 2018-03-02
@Lukasa also I think we want to use this handler in our `HTTP1` example