@normanmaurer 2018-03-02
This will fix https://github.com/apple/swift-nio/issues/43