@weissi 2018-03-03
@Lukasa please run

    git mv IntegrationTests/tests_01_http/SKIP_test_13_http_pipelining.sh  IntegrationTests/tests_01_http/test_13_http_pipelining.sh

to enable the pipelining tes