@jordan2936 2018-03-03
Man I don’t know what that is 


Sent from Yahoo Mail for iPhone


On Friday, March 2, 2018, 6:56 PM, Johannes Weiß <notifications@github.com> wrote:


@Lukasa please run
git mv IntegrationTests/tests_01_http/SKIP_test_13_http_pipelining.sh  IntegrationTests/tests_01_http/test_13_http_pipelining.sh

to enable the pipelining tes

—
You are receiving this because you commented.
Reply to this email directly, view it on GitHub, or mute the thread.