@Lukasa 2018-03-07
@tomerd Seems like the builder is keeping hold of state from #109 which is causing this build to fail. Want to dive into that?