@tomerd 2018-03-08
> @tomerd Seems like the builder is keeping hold of state from #109 which is causing this build to fail. Want to dive into that?

@Lukasa this is fixed now, but this is filling an integration test

```
00:03:49.294 Running test 'test_13_http_pipelining.sh'... FAILURE (1)
```