@Lukasa 2018-03-08
So the failure here is because netcat half-closes the TCP connection when it gets EOF, which confuses the crap out of our test server. I'm going to work on that, but it'll take a few patches because the test server doesn't currently handle keep-alive properly either.