@Lukasa 2018-03-13
Perf results:

#### Without pipeline handler

```
wrk -c 1k -d 1m -t 4 --latency http://localhost:8888
Running 1m test @ http://localhost:8888
  4 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.19ms    3.15ms  56.53ms   84.12%
    Req/Sec    13.83k   842.71    16.24k    80.12%
  Latency Distribution
     50%   17.69ms
     75%   18.28ms
     90%   19.58ms
     99%   22.45ms
  3302121 requests in 1.00m, 160.61MB read
  Socket errors: connect 0, read 918, write 0, timeout 0
Requests/sec:  55004.64
Transfer/sec:      2.68MB
```

With the pipelining handler

```
wrk -c 1k -d 1m -t 4 --latency http://localhost:8888
Running 1m test @ http://localhost:8888
  4 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.84ms    4.89ms 168.78ms   86.04%
    Req/Sec    13.57k     1.02k   17.11k    86.04%
  Latency Distribution
     50%   17.90ms
     75%   18.45ms
     90%   19.69ms
     99%   23.24ms
  3240724 requests in 1.00m, 157.62MB read
  Socket errors: connect 0, read 1131, write 0, timeout 0
Requests/sec:  53972.47
Transfer/sec:      2.63MB
```

With the pipelining handler, and actually pipelining:

```
wrk -c 1k -d 1m -t 4 --latency -s pipeline.lua http://localhost:8888
Running 1m test @ http://localhost:8888
  4 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   317.96ms  210.25ms   1.99s    62.87%
    Req/Sec    47.34k    14.35k   96.72k    66.08%
  Latency Distribution
     50%  291.35ms
     75%  471.95ms
     90%  626.93ms
     99%  824.32ms
  11300061 requests in 1.00m, 549.61MB read
  Socket errors: connect 0, read 2381, write 70, timeout 467
Requests/sec: 188102.51
Transfer/sec:      9.15MB
```

The TL;DR here is that the pipelining handler slows the server down in the non-pipelined mode by a worst-case of about 2% (these tests do no meaningful CPU work in the HTTP server itself, so the performance delta here basically is the worst case).

I'd say that's a small price to pay for correctness. ;)

(BTW, the read errors here are probably because this server closes non-keepalive connections, and that the way it does so could easily lead to TCP RST rather than EOF. We should address that in a future patch.)